rename: walkuseQuery hook 파일이름 변경 #70
